[Feature][Connector-V2] Refactor oss connector
