+++ seatunnel-connectors-v2/connector-file/connector-file-oss/src/main/java/org/apache/seatunnel/connectors/seatunnel/file/oss/sink/OssFileSink.java
@@ -23,16 +23,13 @@ import org.apache.seatunnel.common.config.CheckConfigUtil;
 import org.apache.seatunnel.common.config.CheckResult;
 import org.apache.seatunnel.common.constants.PluginType;
 import org.apache.seatunnel.connectors.seatunnel.file.config.FileSystemType;
-import org.apache.seatunnel.connectors.seatunnel.file.oss.source.config.OssConf;
-import org.apache.seatunnel.connectors.seatunnel.file.oss.source.config.OssSourceConfig;
+import org.apache.seatunnel.connectors.seatunnel.file.oss.config.OssConf;
+import org.apache.seatunnel.connectors.seatunnel.file.oss.config.OssConfig;
 import org.apache.seatunnel.connectors.seatunnel.file.sink.BaseFileSink;
 
 import org.apache.seatunnel.shade.com.typesafe.config.Config;
 
 import com.google.auto.service.AutoService;
-import org.apache.hadoop.fs.aliyun.oss.Constants;
-
-import java.util.HashMap;
 
 @AutoService(SeaTunnelSink.class)
 public class OssFileSink extends BaseFileSink {
@@ -45,17 +42,12 @@ public class OssFileSink extends BaseFileSink {
     public void prepare(Config pluginConfig) throws PrepareFailException {
         super.prepare(pluginConfig);
         CheckResult result = CheckConfigUtil.checkAllExists(pluginConfig,
-                OssSourceConfig.FILE_PATH,
-                OssSourceConfig.BUCKET, OssSourceConfig.ACCESS_KEY,
-                OssSourceConfig.ACCESS_SECRET, OssSourceConfig.BUCKET);
+                OssConfig.FILE_PATH,
+                OssConfig.BUCKET, OssConfig.ACCESS_KEY,
+                OssConfig.ACCESS_SECRET, OssConfig.BUCKET);
         if (!result.isSuccess()) {
             throw new PrepareFailException(getPluginName(), PluginType.SINK, result.getMsg());
         }
-        hadoopConf = new OssConf(pluginConfig.getString(OssSourceConfig.BUCKET));
-        HashMap<String, String> ossOptions = new HashMap<>();
-        ossOptions.put(Constants.ACCESS_KEY_ID, pluginConfig.getString(OssSourceConfig.ACCESS_KEY));
-        ossOptions.put(Constants.ACCESS_KEY_SECRET, pluginConfig.getString(OssSourceConfig.ACCESS_SECRET));
-        ossOptions.put(Constants.ENDPOINT_KEY, pluginConfig.getString(OssSourceConfig.ENDPOINT));
-        hadoopConf.setExtraOptions(ossOptions);
+        hadoopConf = OssConf.buildWithConfig(pluginConfig);
     }
 }
